requirements.txt
python-pptx
python-dotenv
requests
transformers
accelerate
langchain
